requirements.txt
opencv-python
numba
h5py
hdf5plugin
ruff
tqdm
onnxruntime-gpu
torch